# coffee-be/static/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ChainFacility": {
				"description": "Each organization (ChainOrganization) may have several facilities. Facility is considered as a storage space for\nproduct units (ChainProductUnit). Transactions (ChainTransaction) run between facilities and are recorded by\nChainTransaciton objects.",
				"properties": {
					"docType": {
						"type": "string",
						"nullable": true
					},
					"_id": {
						"type": "string",
						"nullable": true
					},
					"_rev": {
						"type": "string",
						"nullable": true
					},
					"created": {
						"type": "string",
						"nullable": true,
						"description": "Timestamp of creation"
					},
					"lastChange": {
						"type": "string",
						"nullable": true,
						"description": "Timestamp of last change"
					},
					"userCreatedId": {
						"type": "string",
						"nullable": true,
						"description": "Id of user that created the document."
					},
					"userChangedId": {
						"type": "string",
						"nullable": true,
						"description": "Id of user that changed the document."
					},
					"name": {
						"type": "string",
						"description": "Facility name"
					},
					"organizationId": {
						"type": "string",
						"description": "Organization id"
					},
					"facilityType": {
						"$ref": "#/components/schemas/ChainFacilityType",
						"description": "Role of the facility in the value chain"
					},
					"isCollectionFacility": {
						"type": "boolean",
						"description": "Facility where cherries are gathered"
					},
					"isPublic": {
						"type": "boolean",
						"nullable": true,
						"description": "Whether anybody can list stock orders from this facility (facility intended for sale)."
					},
					"location": {
						"$ref": "#/components/schemas/ChainLocation",
						"description": "Facility location"
					},
					"organization": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ChainOrganization"
							},
							{
								"type": "number",
								"enum": [
									null
								],
								"nullable": true
							}
						],
						"description": "Organization object. Set automatically on read. Ignored on write (not stored, obtained through organizationId)"
					},
					"semiProducts": {
						"items": {
							"$ref": "#/components/schemas/ChainSemiProduct"
						},
						"type": "array",
						"description": "Allowable semi products"
					},
					"semiProductIds": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "Allowable semi product ids"
					},
					"semiProductPrices": {
						"items": {
							"$ref": "#/components/schemas/ChainSemiProductPrice"
						},
						"type": "array",
						"description": "Suggested semi-product prices"
					}
				},
				"required": [
					"name",
					"organizationId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ChainProductOrder": {
				"description": "Represents a product order. Items in the order are",
				"properties": {
					"docType": {
						"type": "string",
						"nullable": true
					},
					"_id": {
						"type": "string",
						"nullable": true
					},
					"_rev": {
						"type": "string",
						"nullable": true
					},
					"created": {
						"type": "string",
						"nullable": true,
						"description": "Timestamp of creation"
					},
					"lastChange": {
						"type": "string",
						"nullable": true,
						"description": "Timestamp of last change"
					},
					"userCreatedId": {
						"type": "string",
						"nullable": true,
						"description": "Id of user that created the document."
					},
					"userChangedId": {
						"type": "string",
						"nullable": true,
						"description": "Id of user that changed the document."
					},
					"id": {
						"type": "string",
						"description": "Order id."
					},
					"facilityId": {
						"type": "string",
						"description": "Facility to which order is ordered"
					},
					"deliveryDeadline": {
						"type": "string",
						"nullable": true,
						"description": "Order delivery deadline"
					},
					"customerId": {
						"type": "string",
						"description": "Company customer id."
					},
					"requiredwomensOnly": {
						"type": "boolean",
						"description": "Indicator whether only women's coffee is required"
					},
					"requiredGradeId": {
						"type": "string",
						"description": "Grade Id."
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/ChainStockOrder"
						},
						"type": "array",
						"nullable": true,
						"description": "Read only list of ordered items. Used also on creation"
					},
					"facility": {
						"$ref": "#/components/schemas/ChainFacility",
						"description": "Read only facility."
					},
					"customer": {
						"$ref": "#/components/schemas/ChainCompanyCustomer",
						"description": "Read only company customer."
					},
					"requiredGrade": {
						"$ref": "#/components/schemas/ChainGradeAbbreviation",
						"description": "Read only required grade."
					},
					"processingOrders": {
						"items": {
							"$ref": "#/components/schemas/ChainProcessingOrder"
						},
						"type": "array",
						"description": "Processing orders. Insert only. Does not get saved. Used for creating orders."
					},
					"open": {
						"type": "boolean",
						"description": "List of open orders"
					}
				},
				"required": [
					"id",
					"facilityId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedList_ChainProductOrder_": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double",
						"description": "Count of all items satisfying 'paginatable' request."
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/ChainProductOrder"
						},
						"type": "array",
						"description": "Response items."
					},
					"limit": {
						"type": "number",
						"format": "double",
						"description": "Limit got from request"
					},
					"offset": {
						"type": "number",
						"format": "double",
						"description": "Offset got from request"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiValidationErrorDetails": {
				"properties": {
					"className": {
						"type": "string"
					},
					"fieldErrors": {
						"properties": {},
						"additionalProperties": {
							"type": "string"
						},
						"type": "object"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_PaginatedList_ChainProductOrder__": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/PaginatedList_ChainProductOrder_"
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_ProcessingOrderHistory-Array_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ProcessingOrderHistory"
						},
						"type": "array"
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_QuoteRequirementConfirmationsWithMetaData_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/QuoteRequirementConfirmationsWithMetaData"
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ChainProductOrder_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/ChainProductOrder"
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_any_": {
				"properties": {
					"data": {
						"additionalProperties": true
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_PaginatedList_ChainFacility__": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/PaginatedList_ChainFacility_"
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ChainFacility_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/ChainFacility"
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "coffee-be",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/chain-api/data/order/facility/{facilityId}/list-open": {
			"get": {
				"operationId": "ListOpenOrdersForFacility",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainProductOrder__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Order"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "facilityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "openOnly",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/data/order/organization/{organizationId}/list-open": {
			"get": {
				"operationId": "ListOpenOrdersForOrganization",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainProductOrder__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Order"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "organizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "openOnly",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/data/order/aggregates-for-order/{orderId}": {
			"get": {
				"operationId": "GetAggregatesForOrder",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ProcessingOrderHistory-Array_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Order"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/chain-api/data/order/quote-requrements-verify-for-order/{orderId}": {
			"get": {
				"operationId": "GetQuoteRequirementsVerification",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_QuoteRequirementConfirmationsWithMetaData_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Order"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/chain-api/data/order/{dbId}": {
			"get": {
				"operationId": "GetOrder",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ChainProductOrder_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Fetches an order by its _id.",
				"tags": [
					"Order"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "dbId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/chain-api/data/order": {
			"post": {
				"operationId": "PostOrder",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Inserts or updates an order. When inserting fields _id, _rev and docType should not be present.",
				"tags": [
					"Order"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainProductOrder"
							}
						}
					}
				}
			}
		},
		"/chain-api/data/order/delete": {
			"post": {
				"operationId": "DeleteOrder",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Deletes order.",
				"tags": [
					"Order"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainProductOrder"
							}
						}
					}
				}
			}
		},
...
		"/chain-api/data/facility/list": {
			"get": {
				"operationId": "ListFacilities",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainFacility__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Paginated list of facilities.",
				"tags": [
					"Facility"
				],
				"security": [],
				"parameters": [
					{
						"description": "sort order ASC or DESC",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"description": "query limit",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "query offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/data/facility/{dbId}": {
			"get": {
				"operationId": "GetFacilityById",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ChainFacility_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Fetches a facility by its _id.",
				"tags": [
					"Facility"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "dbId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/chain-api/data/facility": {
			"post": {
				"operationId": "PostFacility",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Inserts or updates a facility. When inserting fields _id, _rev and docType should not be present.",
				"tags": [
					"Facility"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainFacility"
							}
						}
					}
				}
			}
		},
		"/chain-api/data/facility/list/organization/{organizationId}": {
			"get": {
				"operationId": "ListFacilitiesForOrganization",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainFacility__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Paginated list of facilities for a given organization",
				"tags": [
					"Facility"
				],
				"security": [],
				"parameters": [
					{
						"description": "_id of a given organization",
						"in": "path",
						"name": "organizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "sort order ASC or DESC",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"description": "query limit",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "query offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/data/facility/list/organization/{organizationId}/semi-product/{semiProductId}": {
			"get": {
				"operationId": "ListFacilitiesForOrganizationAndSemiProduct",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainFacility__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Paginated list of facilities for a given organization and semiProductId",
				"tags": [
					"Facility"
				],
				"security": [],
				"parameters": [
					{
						"description": "_id of a given organization",
						"in": "path",
						"name": "organizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "_id of desired semi-product",
						"in": "path",
						"name": "semiProductId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "sort order ASC or DESC",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"description": "query limit",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "query offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/data/facility/list/organization/{organizationId}/semi-product/{semiProductId}/selling": {
			"get": {
				"operationId": "ListSellingFacilitiesForOrganizationAndSemiProduct",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainFacility__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Paginated list of sellig facilities for a given organization and semiProductId",
				"tags": [
					"Facility"
				],
				"security": [],
				"parameters": [
					{
						"description": "_id of a given organization",
						"in": "path",
						"name": "organizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "_id of desired semi-product",
						"in": "path",
						"name": "semiProductId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "sort order ASC or DESC",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"description": "query limit",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "query offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/data/facility/list/collecting/organization/{organizationId}": {
			"get": {
				"operationId": "ListOfCollectingFacilitiesForOrganization",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainFacility__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Paginated list of COLLECTING facilities for a given organization",
				"tags": [
					"Facility"
				],
				"security": [],
				"parameters": [
					{
						"description": "_id of a given organization",
						"in": "path",
						"name": "organizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "sort order ASC or DESC",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"description": "query limit",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "query offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/data/facility/delete": {
			"post": {
				"operationId": "DeleteFacility",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Deletes facility.",
				"tags": [
					"Facility"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainFacility"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}